% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstate_functions.R
\docType{package}
\name{appstate_googleAuthR}
\alias{appstate_googleAuthR}
\alias{appstate_googleAuthR-package}
\title{Google App State API
The Google App State API.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:36:04
filename: /Users/mark/dev/R/autoGoogleAPI/googleappstatev1.auto/R/appstate_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/appstate
}
}

