% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstate_objects.R
\name{GetResponse}
\alias{GetResponse}
\title{Google App State API Objects 
The Google App State API.}
\usage{
GetResponse(currentStateVersion = NULL, data = NULL, stateKey = NULL)
}
\arguments{
\item{currentStateVersion}{The current app state version}

\item{data}{The requested data}

\item{stateKey}{The key for the data}
}
\value{
GetResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:36:04
filename: /Users/mark/dev/R/autoGoogleAPI/googleappstatev1.auto/R/appstate_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
GetResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an app state resource.
}

