% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstate_objects.R
\name{ListResponse}
\alias{ListResponse}
\title{ListResponse Object}
\usage{
ListResponse(items = NULL, maximumKeyCount = NULL)
}
\arguments{
\item{items}{The app state data}

\item{maximumKeyCount}{The maximum number of keys allowed for this user}
}
\value{
ListResponse object
}
\description{
ListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template to convert a list-response for app state.
}

