% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstate_functions.R
\name{states.clear}
\alias{states.clear}
\title{Clears (sets to empty) the data for the passed key if and only if the passed version matches the currently stored version. This method results in a conflict error on version mismatch.}
\usage{
states.clear(stateKey, currentDataVersion = NULL)
}
\arguments{
\item{stateKey}{The key for the data to be retrieved}

\item{currentDataVersion}{The version of the data to be cleared}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/appstate
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/appstate)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/web/api/states}{Google Documentation}
}

