% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstate_functions.R
\name{states.delete}
\alias{states.delete}
\title{Deletes a key and the data associated with it. The key is removed and no longer counts against the key quota. Note that since this method is not safe in the face of concurrent modifications, it should only be used for development and testing purposes. Invoking this method in shipping code can result in data loss and data corruption.}
\usage{
states.delete(stateKey)
}
\arguments{
\item{stateKey}{The key for the data to be retrieved}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/appstate
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/appstate)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/web/api/states}{Google Documentation}
}

