% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstate_functions.R
\name{states.get}
\alias{states.get}
\title{Retrieves the data corresponding to the passed key. If the key does not exist on the server, an HTTP 404 will be returned.}
\usage{
states.get(stateKey)
}
\arguments{
\item{stateKey}{The key for the data to be retrieved}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/appstate
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/appstate)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/web/api/states}{Google Documentation}
}
