% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstate_objects.R
\name{UpdateRequest}
\alias{UpdateRequest}
\title{UpdateRequest Object}
\usage{
UpdateRequest(data = NULL)
}
\arguments{
\item{data}{The new app state data that your application is trying to update with}
}
\value{
UpdateRequest object
}
\description{
UpdateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a requests which update app state
}
\seealso{
Other UpdateRequest functions: \code{\link{states.update}}
}
