% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstate_objects.R
\name{WriteResult}
\alias{WriteResult}
\title{WriteResult Object}
\usage{
WriteResult(currentStateVersion = NULL, stateKey = NULL)
}
\arguments{
\item{currentStateVersion}{The version of the data for this key on the server}

\item{stateKey}{The written key}
}
\value{
WriteResult object
}
\description{
WriteResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an app state write result.
}

